Add 'Top Gainers' widget type
