+++ app/src/main/res/xml/appwidget_provider_info.xml
@@ -3,11 +3,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:configure="io.horizontalsystems.bankwallet.widgets.MarketWidgetConfigurationActivity"
     android:minWidth="250dp"
-    android:minHeight="150dp"
+    android:minHeight="230dp"
     android:previewImage="@drawable/widget_preview"
     android:resizeMode="horizontal|vertical"
     android:targetCellWidth="5"
-    android:targetCellHeight="4"
+    android:targetCellHeight="5"
     android:updatePeriodMillis="1800000"
     android:widgetCategory="home_screen"
     android:widgetFeatures="reconfigurable"
